Refactor auth logic to genre grid container and add per-component error catching
