+++ cypress.config.js
@@ -3,6 +3,6 @@ const { defineConfig } = require('cypress');
 module.exports = defineConfig({
   projectId: '5te4qf',
   e2e: {
-    baseUrl: 'http://localhost:3000', // Update with your app's URL
+    baseUrl: 'http://localhost:3000', 
   },
 });
