+++ src/index.css
@@ -13,14 +13,13 @@ code {
 }
 
 .error-fallback {
-  display: flex;
   flex-direction: column;
   align-items: center;
   justify-content: center;
-  height: 100vh;
   text-align: center;
-  color: #ffffff;
+  color: #FFFFFF;
   background-color: #121212;
+  padding-top: 50px;
 }
 
 .error-title {
